LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/flow-content-0/dialog-display-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_equals: expected "block" but got "inline"

FAIL dialog: display Error: assert_equals: expected "block" but got "inline"
